GFN-IoT-Project.Collecto/requirements.txt
pyserial
smbus
RPi.bme280 
smbus2